Feat: Add some minor fixes
Styles are fixed, and components are separated into files. Comments are added.
