+++ src/documentRenderers/richtext/RichTextRenderer.module.css
@@ -54,6 +54,7 @@
 }
 
 .rich-text-option {
+  flex-grow: 1;
   list-style: none;
   font-size: xx-large;
   padding: 1em;
